目的：ビルド時の警告を全て解消し、Easingクラスの不具合(InOut系のelse抜け)も修正した.
変更点
PMDActor / PMXActor / PMXRenderer / XActor
・float/size_t等からUINTへの暗黙の縮小変換警告(C4244/C4267)が出ていた箇所に明示的なstatic_castを追加(値は変わらず、警告のみ解消).

Easing.inl
・M_PI(double)をそのままfloat演算に混ぜていたためdouble→float縮小変換警告(C4244)が多数出ていた. float版定数PI_Fを追加して置き換えた.
・std::pow(2, ...)の第1引数がint(→doubleオーバーロードに解決)だったのをstd::pow(2.0f, ...)に変更.
・ついでに発見した不具合: InOutQuad/InOutCubic/InOutQuart/InOutQuint/InOutExpo/InOutCirc/InOutBack/InElastic/OutElasticが、tによる条件分岐の結果をelse/returnせず直後の無条件文で毎回上書きしていたため、前半(t<0.5等)の値が常に後半の式で潰されていた. 各分岐にreturnを追加して修正.

VMDLoader.cpp
・catch節のeが未使用(C4101)だった点を修正済み(e.what()をエラーメッセージへ含めるよう変更).

DirectX12.vcxproj
・PostBuildEventのxcopyに/Dを追加. 変更の無いファイルは毎回コピーし直さないようにし、ビルドのたびに時間がかかっていたのを解消(検証: 変更無しビルドで所要時間が約1秒、コピー0件に).

影響範囲（他ファイルにまたがるとき）
なし. Debug/Releaseとも警告0件・エラー0件でビルド確認済み.

+++ DirectX12/SourceCode/10_Ggraphic/PMD/PMDActor.cpp
@@ -471,7 +471,7 @@ void PMDActor::PlayAnimation()
 void PMDActor::MotionUpdate()
 {
 	auto elapsedTime = timeGetTime() - m_StartTime;
-	unsigned int frameNo = 30 * (elapsedTime / 1000.0f);
+	unsigned int frameNo = static_cast<unsigned int>(30 * (elapsedTime / 1000.0f));
 
 
 	//行列をクリア(しないと前フレームのポーズが重ね掛けされてモデルが崩れる)
